refac(DPLAN-18226): create a composable for transforming the EDT filter to API 3.0
